+++ src/startup.rs
@@ -1,0 +1,15 @@
+use crate::routes;
+use actix_web::dev::Server;
+use actix_web::{web, App, HttpServer};
+use std::net::TcpListener;
+
+pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
+    let server = HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(routes::health_check))
+            .route("/subscriptions", web::post().to(routes::subscribe))
+    })
+    .listen(listener)?
+    .run();
+    Ok(server)
+}
