+++ tests/health_check.rs
@@ -20,7 +20,7 @@ fn spawn_app() -> String {
     // port 0 is special OS wise - OS will scan and give us the first open port
     let listener = TcpListener::bind("127.0.0.1:0").expect("faield to bind to random port");
     let port = listener.local_addr().unwrap().port();
-    let server = zero2prod::run(listener).expect("Failed to bind address");
+    let server = zero2prod::startup::run(listener).expect("Failed to bind address");
 
     let _ = tokio::spawn(server);
     format!("http://127.0.0.1:{}", port)
